add UserMutation
