+++ App1/GraphQLTypes/RootMutation.cs
@@ -12,55 +12,11 @@ namespace IOprojekt.GraphQLTypes
     public class RootMutation : ObjectGraphType
     {
         private readonly IRepository<User> users;
-        public RootMutation(IDbContext context, IOptions<MongoSettings> conf)
+        public RootMutation(IDbContext context)
         {
 
-            if ( context != null)
-                users = context.Users;
-
-            Field<UserType>("addUser",
-                arguments: new QueryArguments
-                {
-                    new QueryArgument<InputUserType>() { Name = "user" }
-                },
-                resolve: context =>
-                {
-                    var user = context.GetArgument<User>("user");
-                    return users.Add(user);
-                }
-             );
-
-            Field<IntGraphType>("deleteUser",
-                arguments: new QueryArguments
-                {
-                     new QueryArgument<IntGraphType>() { Name = "userId" }
-                },
-                resolve: context =>
-                {
-                    var id = context.GetArgument<int>("userId");
-                    var builder = Builders<User>.Filter;
-                    var filter = builder.Eq(user => user.Id, id);
-                    return users.Delete(filter);
-                }
-             );
-
-            Field<UserType>("updateUser",
-               arguments: new QueryArguments
-               {
-                     new QueryArgument<InputUserType>() { Name = "user" }
-               },
-               resolve: context =>
-               {
-                   var user = context.GetArgument<User>("user");
-                   var builder = Builders<User>.Filter;
-                   var filter = builder.Eq(user => user.Id, user.Id);
-                   var update = Builders<User>.Update
-                                        .Set("firstName", user.FirstName)
-                                        .Set("lastName", user.LastName);
-
-                   return users.Update(filter, update);
-               }
-            );
+            Field<UserMutation>("users", resolve: context => new { });
+           // Field<PostQuery>("postsQuery", resolve: context => new { });
 
         }
     }
